+++ lib/ftpd/auth_levels.rb
@@ -1,0 +1,9 @@
+module Ftpd
+
+  # Authorization levels for FtpServer#auth_level
+
+  AUTH_USER = 0
+  AUTH_PASSWORD = 1
+  AUTH_ACCOUNT = 2
+
+end